+++ src/main/java/com/tallerwebi/infraestructura/RepositorioUsuarioImpl.java
@@ -75,7 +75,7 @@ public class RepositorioUsuarioImpl implements RepositorioUsuario {
     return existente != null;
   }
 
-  public Usuario buscarUsuarioPorId(Integer id) {
+  public Usuario buscarUsuarioPorId(Long id) {
     return sessionFactory
       .getCurrentSession()
       .createQuery("from Usuario where id = :id", Usuario.class)
